Investigate having -0 dollar funds in stock purchases, changed to BigDecimal instead of double

+++ src/main/java/com/jeno/fantasyleague/util/DecimalUtil.java
@@ -1,5 +1,6 @@
 package com.jeno.fantasyleague.util;
 
+import java.math.BigDecimal;
 import java.text.DecimalFormat;
 import java.text.DecimalFormatSymbols;
 import java.text.NumberFormat;
@@ -7,7 +8,7 @@ import java.util.Locale;
 
 public class DecimalUtil {
 
-	public static String getTwoDecimalsThousandSeperator(double value) {
+	public static String getTwoDecimalsThousandSeperator(BigDecimal value) {
 		DecimalFormat formatter = (DecimalFormat) NumberFormat.getInstance(Locale.US);
 		DecimalFormatSymbols symbols = formatter.getDecimalFormatSymbols();
